Moved reference collection stuff to trait
